Rework recipe ingredients to percent.
Rework recipe ingredients to only specify the percentage instead of the
amount. Add field to the recipe to specify total amount of dry
ingredients and use that to calculate base amounts for all ingredients.

Also fix bug with which ingredients show up in Sanity when selecting
ingredients for a recipe.